feat: consolidate login and signup into a single auth page with tab toggle
Created a unified AuthPage component that handles both login and signup
flows via a segmented tab control. Google OAuth moved above the email form
as the primary CTA since most users prefer it.

Key changes:
- New AuthPage.tsx with shared form logic, validation, and error handling
- LoginPage.tsx and SignupPage.tsx reduced to thin wrappers (3 lines each)
- Removed decorative gradient backgrounds for a cleaner, more premium look
- Brand shown as a compact icon instead of large logo image
- Input fields use rounded-lg (8px) instead of rounded-xl (16px) for
  tighter form feel
- Tab toggle lets users switch between login/signup without page navigation

Addresses: #779
